+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Thunderdome Changelog
 
 ## Unreleased Changes
+* Fix `Arena::iter_mut` to return mutable references.
 
 ## 0.3.0 (2020-10-16)
 * Added `Arena::invalidate` for invalidating indices on-demand, as a faster remove-followed-by-reinsert.