+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Thunderdome Changelog
 
 ## Unreleased Changes
+* Added `Arena::retain` for conveniently removing entries which do not satisfy a given predicate.
 * Fix `Arena::iter_mut` to return mutable references.
 
 ## 0.3.0 (2020-10-16)